requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
dnspython==2.1.0
pycparser==2.20
pymongo==3.11.4
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1